HW2.7
