+++ src/main/java/pro/sky/hw27/controller/EmployeeController.java
@@ -1,14 +1,13 @@
-package pro.sky.HW27.controller;
+package pro.sky.hw27.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import pro.sky.HW27.Employee;
-import pro.sky.HW27.employeeService.EmployeeService;
+import pro.sky.hw27.Employee;
+import pro.sky.hw27.employeeService.EmployeeService;
 
 
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -20,11 +19,6 @@ public class EmployeeController {
         this.employeeService = employeeService;
     }
 
-    @GetMapping("/add")
-    public Employee addEmployee(@RequestParam("name") String name) {
-        return employeeService.addEmployee(name);
-
-    }
 
     @GetMapping("/find")
     public Employee findEmployee(@RequestParam("name") String name) {
@@ -38,30 +32,9 @@ public class EmployeeController {
 
     }
 
-    @GetMapping("/all")
-    public Map<String,Employee> getAll(@RequestParam("id") int id) {
-        return employeeService.getAll(id);
-    }
-    @GetMapping("/all-by-dep")
-    public Map<String,Employee> getAllByDep() {
-        return employeeService.getAllByDep();
-    }
-
-
-
-    @GetMapping("/min-salary")
-    public Employee minSalary(@RequestParam("id") int id ) {
-        return employeeService.minSalary(id);
-
-    }
-    @GetMapping("/max-salary")
-    public Employee maxSalary(@RequestParam("id") int id ) {
-        return employeeService.maxSalary(id);
-
-    }
-    @GetMapping("/add2")
+    @GetMapping("/add")
     public Employee addEmployee(@RequestParam("name") String name,@RequestParam("id") int id,@RequestParam("salary") int salary) {
-        return employeeService.addEmployee2(name,id,salary);
+        return employeeService.addEmployee(name,id,salary);
 
     }
