+++ src/main/java/pro/sky/hw27/employeeService/EmployeeService.java
@@ -1,0 +1,76 @@
+package pro.sky.hw27.employeeService;
+
+import org.springframework.stereotype.Service;
+import pro.sky.hw27.Employee;
+import pro.sky.hw27.exeption.EmployeeAlreadyAddedExeption;
+import pro.sky.hw27.exeption.EmployeeNotFoundExeption;
+
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+
+@Service
+public class EmployeeService {
+    private static final int LIMIT = 10;
+    private final Map<String, Employee> employees = new HashMap<>();
+
+    public Employee addEmployee(String name, int id, int salary) {
+        Employee employee = new Employee(name, id, salary);
+        if (employees.containsKey(name)){
+            throw new EmployeeAlreadyAddedExeption();
+        }
+        else{
+            employees.put(name, employee);
+            return employee;
+        }
+
+    }
+
+
+    public Employee findEmployee(String name) {
+        Employee employee = new Employee(name);
+        if (employees.containsKey(name)) {
+            return employee;
+        } else {
+            throw new EmployeeNotFoundExeption();
+        }
+
+    }
+
+    public Employee removeEmployee(String name) {
+        Employee employee = new Employee(name);
+        if (!employees.containsKey(name)) {
+            throw new EmployeeNotFoundExeption();
+        }
+        employees.remove(name);
+        return employee;
+    }
+
+    public List<Employee> getAllByDep(int id) {
+        return employees.values().stream().filter(e -> e.getEmployeeDepartment() == id).collect(Collectors.toList());
+    }
+
+    public Map<Integer, List<Employee>> getAll() {
+        return employees.values().stream()
+                .collect(Collectors.groupingBy(Employee::getEmployeeDepartment));
+    }
+
+    public Employee minSalary(int i) {
+        return employees.values().stream().filter(e -> e.getEmployeeDepartment() == i).min((e1, e2) -> {
+            return e1.getEmployeeSalary() - e2.getEmployeeSalary();
+        }).orElseThrow(() ->  new EmployeeNotFoundExeption());
+
+    }
+
+    public Employee maxSalary(int i) {
+        return employees.values().stream().filter(e -> e.getEmployeeDepartment() == i).max((e1, e2) -> {
+            return e1.getEmployeeSalary() - e2.getEmployeeSalary();
+        }).orElseThrow(() -> new EmployeeNotFoundExeption());
+
+    }
+
+
+}
