+++ src/main/java/pro/sky/hw27/exeption/EmployeeAlreadyAddedExeption.java
@@ -1,4 +1,4 @@
-package pro.sky.HW27.exeption;
+package pro.sky.hw27.exeption;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
